model/requirements.txt
spacy==3.7.2
skillNer==1.0.3
pdfplumber==0.11.0
python-docx==1.1.2
streamlit==1.39.0